Move PageTitle out of holocene and use params for title
