+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@temporalio/ui",
-  "version": "2.1.69",
+  "version": "2.1.70",
   "type": "module",
   "description": "Temporal.io UI",
   "keywords": [
