+++ src/routes/login/index@login.svelte
@@ -28,12 +28,14 @@
   import NavigationHeader from '$lib/components/navigation-header.svelte';
   import HamburgerHeader from '$lib/components/hamburger-header.svelte';
   import { publicPath } from '$lib/utilities/get-public-path';
+  import PageTitle from '$lib/components/page-title.svelte';
 
   const error = $page.url.searchParams.get('error');
 
   export let settings: Settings;
 </script>
 
+<PageTitle title="Login" url={$page.url.href} />
 <NavigationHeader href="{publicPath}/" user={undefined} />
 <HamburgerHeader href="{publicPath}/" user={undefined} />
 <section class="my-[20vh] text-center">
