+++ src/routes/namespaces/[namespace]/schedules/new.svelte
@@ -2,11 +2,10 @@
   import { page } from '$app/stores';
 
   import SchedulesNew from '$lib/pages/schedules-new.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
+
+  const namespace = $page.params.namespace;
 </script>
 
-<PageTitle
-  title={`Create Schedule | ${$page.params.namespace}`}
-  url={$page.url.href}
-/>
+<PageTitle title={`Create Schedule | ${namespace}`} url={$page.url.href} />
 <SchedulesNew />
