+++ src/routes/namespaces/[namespace]/task-queues/[queue].svelte
@@ -2,8 +2,10 @@
   import { page } from '$app/stores';
 
   import TaskQueueWorkers from '$lib/pages/task-queue-workers.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
+
+  const queue = $page.params.queue;
 </script>
 
-<PageTitle title={`Task Queue | ${$page.params.queue}`} url={$page.url.href} />
+<PageTitle title={`Task Queue | ${queue}`} url={$page.url.href} />
 <TaskQueueWorkers />
