+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/__layout.svelte
@@ -3,13 +3,12 @@
   import { workflowRun } from '$lib/stores/workflow-run';
 
   import WorkflowHistoryLayout from '$lib/layouts/workflow-history-layout.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
+
+  const workflow = $page.params.workflow;
 </script>
 
-<PageTitle
-  title={`Workflow History | ${$workflowRun?.workflow?.id}`}
-  url={$page.url.href}
-/>
+<PageTitle title={`Workflow History | ${workflow}`} url={$page.url.href} />
 <WorkflowHistoryLayout>
   <!-- <svelte:fragment slot="timeline">
     <EventHistoryTimelineContainer />
