+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/pending-activities.svelte
@@ -1,13 +1,11 @@
 <script lang="ts">
   import { page } from '$app/stores';
-  import { workflowRun } from '$lib/stores/workflow-run';
 
   import WorkflowPendingActivities from '$lib/pages/workflow-pending-activities.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
+
+  const workflow = $page.params.workflow;
 </script>
 
-<PageTitle
-  title={`Pending Activities | ${$workflowRun?.workflow?.id}`}
-  url={$page.url.href}
-/>
+<PageTitle title={`Pending Activities | ${workflow}`} url={$page.url.href} />
 <WorkflowPendingActivities />
