+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/workers.svelte
@@ -1,13 +1,11 @@
 <script lang="ts">
   import { page } from '$app/stores';
-  import { workflowRun } from '$lib/stores/workflow-run';
 
   import WorkflowWorkers from '$lib/pages/workflow-workers.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
+
+  const workflow = $page.params.workflow;
 </script>
 
-<PageTitle
-  title={`Workers | ${$workflowRun?.workflow?.id}`}
-  url={$page.url.href}
-/>
+<PageTitle title={`Workers | ${workflow}`} url={$page.url.href} />
 <WorkflowWorkers />
