+++ src/routes/namespaces/[namespace]/workflows/index@root.svelte
@@ -2,11 +2,11 @@
   import { page } from '$app/stores';
 
   import Workflows from '$lib/pages/workflows.svelte';
-  import PageTitle from '$lib/holocene/page-title.svelte';
+  import PageTitle from '$lib/components/page-title.svelte';
 </script>
 
 <PageTitle
-  title={`Workflows | ${$page.params?.namespace}`}
+  title={`Workflows | ${$page.params.namespace}`}
   url={$page.url.href}
 />
 <Workflows />
